refactor: ClassRoom 모듈 분리
